Separate mixins and interfaces and application group interfaces
