+++ src/beamlime/applications/__init__.py
@@ -1,0 +1,10 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2023 Scipp contributors (https://github.com/scipp)
+# flake8: noqa F401
+
+from .groups import MAX_INSTANCE_NUMBER, BeamlimeApplicationInstanceGroup
+from .interfaces import (
+    MAX_INSTANCE_PAUSED,
+    BeamlimeApplicationInterface,
+    BeamlimeDataReductionInterface,
+)
